boards: nrf53: remove redundant cpunet initialization code
This is now handled at SoC level, avoiding code duplicates.
