+++ boards/ezurio/bl5340_dvk/Kconfig
@@ -21,26 +21,9 @@ config BOARD_ENABLE_DCDC_HV
 	select SOC_DCDC_NRF53X_HV
 	default y
 
-config BOARD_ENABLE_CPUNET
-	bool "NRF53 Network MCU"
-	select SOC_NRF_GPIO_FORWARDER_FOR_NRF5340 if \
-		$(dt_compat_enabled,$(DT_COMPAT_NORDIC_NRF_GPIO_FORWARDER))
-	help
-	  This option enables releasing the Network 'force off' signal, which
-	  as a consequence will power up the Network MCU during system boot.
-	  Additionally, the option allocates GPIO pins that will be used by UARTE
-	  of the Network MCU.
-	  Note: GPIO pin allocation can only be configured by the secure Application
-	  MCU firmware, so when this option is used with the non-secure version of
-	  the board, the application needs to take into consideration, that the
-	  secure firmware image must already have configured GPIO allocation for the
-	  Network MCU.
-	default y if (BT || NRF_802154_SER_HOST)
-
 config DOMAIN_CPUNET_BOARD
 	string
 	default "bl5340_dvk/nrf5340/cpunet"
-	depends on BOARD_ENABLE_CPUNET
 	help
 	  The board which will be used for CPUNET domain when creating a multi
 	  image application where one or more images should be located on
