+++ boards/nordic/thingy53/board.c
@@ -6,11 +6,6 @@
 
 #include <zephyr/init.h>
 #include <zephyr/kernel.h>
-#include <zephyr/logging/log.h>
-#include <soc.h>
-#include <hal/nrf_reset.h>
-
-LOG_MODULE_REGISTER(thingy53_board_init);
 
 /* Initialization chain of Thingy:53 board requires some delays before on board sensors
  * could be accessed after power up. In particular bme680 and bmm150 sensors require,
@@ -34,52 +29,20 @@ BUILD_ASSERT(CONFIG_THINGY53_INIT_PRIORITY < CONFIG_SENSOR_INIT_PRIORITY,
 	"CONFIG_THINGY53_INIT_PRIORITY must be less than CONFIG_SENSOR_INIT_PRIORITY");
 #endif
 
-static void enable_cpunet(void)
-{
-#if !defined(CONFIG_TRUSTED_EXECUTION_NONSECURE)
-	/* Retain nRF5340 Network MCU in Secure domain (bus
-	 * accesses by Network MCU will have Secure attribute set).
-	 */
-	NRF_SPU->EXTDOMAIN[0].PERM = 1 << 4;
-#endif /* !CONFIG_TRUSTED_EXECUTION_NONSECURE */
-
-#if !defined(CONFIG_TRUSTED_EXECUTION_SECURE)
-	/*
-	 * Building Zephyr with CONFIG_TRUSTED_EXECUTION_SECURE=y implies
-	 * building also a Non-Secure image. The Non-Secure image will, in
-	 * this case do the remainder of actions to properly configure and
-	 * boot the Network MCU.
-	 */
-
-	/* Release the Network MCU, 'Release force off signal' */
-	nrf_reset_network_force_off(NRF_RESET, false);
-
-	LOG_DBG("Network MCU released.");
-#endif /* !CONFIG_TRUSTED_EXECUTION_SECURE */
-}
-
+#if !defined(CONFIG_TRUSTED_EXECUTION_SECURE) && defined(CONFIG_SENSOR)
 static int setup(void)
 {
-
-#if !defined(CONFIG_TRUSTED_EXECUTION_SECURE)
-	if (IS_ENABLED(CONFIG_SENSOR)) {
-		/* Initialization chain of Thingy:53 board requires some delays before on board
-		 * sensors could be accessed after power up. In particular bme680 and bmm150
-		 * sensors require, 2ms and 1ms power on delay respectively. In order not to sum
-		 * delays, common delay is introduced in the board start up file. This code is
-		 * executed after sensors are powered up and before their initialization.
-		 * It's ensured by build asserts at the beginning of this file.
-		 */
-		k_msleep(2);
-	}
-
-#endif /* !CONFIG_TRUSTED_EXECUTION_SECURE */
-
-	if (IS_ENABLED(CONFIG_BOARD_ENABLE_CPUNET)) {
-		enable_cpunet();
-	}
+	/* Initialization chain of Thingy:53 board requires some delays before on board
+	 * sensors could be accessed after power up. In particular bme680 and bmm150
+	 * sensors require, 2ms and 1ms power on delay respectively. In order not to sum
+	 * delays, common delay is introduced in the board start up file. This code is
+	 * executed after sensors are powered up and before their initialization.
+	 * It's ensured by build asserts at the beginning of this file.
+	 */
+	k_msleep(2);
 
 	return 0;
 }
 
 SYS_INIT(setup, POST_KERNEL, CONFIG_THINGY53_INIT_PRIORITY);
+#endif /* !CONFIG_TRUSTED_EXECUTION_SECURE && CONFIG_SENSOR */
