+++ boards/panasonic/pan1783/Kconfig
@@ -20,23 +20,11 @@ config BOARD_ENABLE_DCDC_HV
 	select SOC_DCDC_NRF53X_HV
 	default y
 
-config BOARD_ENABLE_CPUNET
-	bool "NRF53 Network MCU"
-	select SOC_NRF_GPIO_FORWARDER_FOR_NRF5340 if \
-		$(dt_compat_enabled,$(DT_COMPAT_NORDIC_NRF_GPIO_FORWARDER))
-	help
-	  This option enables releasing the Network 'force off' signal, which
-	  as a consequence will power up the Network MCU during system boot.
-	  Additionally, the option allocates GPIO pins that will be used by UARTE
-	  of the Network MCU.
-	default y if (BT || NRF_802154_SER_HOST)
-
 config DOMAIN_CPUNET_BOARD
 	string
 	default "pan1783_evb/nrf5340/cpunet" if BOARD_PAN1783_EVB_NRF5340_CPUAPP
 	default "pan1783a_evb/nrf5340/cpunet" if BOARD_PAN1783A_EVB_NRF5340_CPUAPP
 	default "pan1783a_pa_evb/nrf5340/cpunet" if BOARD_PAN1783A_PA_EVB_NRF5340_CPUAPP
-	depends on BOARD_ENABLE_CPUNET
 	help
 	  The board which will be used for CPUNET domain when creating a multi
 	  image application where one or more images should be located on
